requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
email-validator==2.0.0.post2
httpx==0.24.1
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.7
openai>=1.0.0
anthropic==0.18.1
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1 